% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/draw.R
\docType{data}
\name{crude}
\alias{crude}
\title{Data on international trade}
\format{Non-symmetric bunary matrix, trade goes from rows to columns.}
\source{
\url{http://moreno.ss.uci.edu/data.html#trade}
}
\usage{
data(crude)
}
\description{
24 selected representative countries exchanging crude materials
}
\keyword{datasets}

