% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/draw.R
\name{draw}
\alias{draw}
\title{draw a clustered graph}
\usage{
draw(A = matrix(), groups = list(),
  colors = terrain.colors(length(groups)), main = "")
}
\arguments{
\item{A}{symmetric adjacency matrix}

\item{groups}{list where each element denotes the cluster, possibly \code{lclust(A, n)}}

\item{colors}{palette which consist of at least as much as number of groups colors, default is
\code{terrain.colors(legth(group))}}
}
\description{
Function \code{draw} employs the output of \code{lclust(A, n)} or any arbitrary list of groups
to plot a nice clustered graph.
}
\examples{
A <- Crude
draw(A, lclust(A, 2))
B <- matrix(c(0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0), 4, 4)
groups <- list(c(1, 3), c(2, 4))
draw(B, groups)
}

