% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/draw.R
\name{draw}
\alias{draw}
\title{draw a clustered graph}
\usage{
draw(A = matrix(), groups = list(),
  colors = terrain.colors(length(groups)), main = "")
}
\arguments{
\item{A}{symmetric adjacency matrix}

\item{groups}{list where each element denotes the cluster, possibly the result of \code{lclust(A, n)}}

\item{colors}{palette which consist of at least \code{length(groups)} colors, default is
\code{terrain.colors(legth(group))}}
}
\description{
Function \code{draw} employs the output of \code{lclust(A, n)} or any arbitrary grouping
to plot a nice clustered graph.
}
\examples{
draw(A, )
}

